backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic==1.17.2
pydantic-settings==2.1.0
python-dotenv
asyncpg==0.31.0
loguru==0.7.2
python-docx
openpyxl
PyJWT
passlib[bcrypt]